Make the “add” script more robust
It turns out that

    xsel --output | xsel --append --secondary

suffers a segmentation fault if secondary is empty.  Thus secondary does
not receive the contents from primary.  Therefore we need to check if
secondary is empty and if so “overwrite” secondary with `--input`.  If
not, we use `--append`.

All of the other scripts which use `--append` have the same problem.  So
I will need to fix them as well.