Remove most scripts
Initially I planned to make a lot of little scripts which could be
combined.  Now I think that was premature.  I should first get some
experience using some of these scripts.  And I have an idea of a
workflow I could use.  It’s not ideal and it will most probably be
improved upon, but it will be easier to improve on it with some
experience compared to just trying to get it right straight away.

For now I only plan to use the following functionalities:

• Add the contents of primary to secondary (`cs-primary-to-secondary`).
• Pop the last value inserted into secondary and put it in
  clipboard (`cs-stack-secondary-pop-to-clipboard`).